pyproject.toml
[project]
name = "eleventools"
version = "0.3.0"
description = "A comprehensive toolbox for ElevenLabs, providing a user-friendly interface for text-to-speech generation with advanced features and bulk processing capabilities"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Manuel Thomsen", email = "manuel@example.com"}
]
keywords = [
    "elevenlabs",
    "text-to-speech",
    "tts",
    "voice-generation",
    "streamlit",
    "audio",
    "bulk-processing",
    "openrouter",
    "ai",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
[project.urls]
Homepage = "https://github.com/SkelegonDK/ElevenTools"
Repository = "https://github.com/SkelegonDK/ElevenTools"
Issues = "https://github.com/SkelegonDK/ElevenTools/issues"
Documentation = "https://github.com/SkelegonDK/ElevenTools#readme"
dependencies = [
    "streamlit>=1.32.0",
    "black>=24.0.0",
    "openai>=1.12.0",
    "urllib3>=2.2.0",
    "pytest>=8.0.0",
    "pyperclip>=1.8.2",
    "elevenlabs>=0.3.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "pytest-cov>=4.1.0",
    "pandas>=2.2.3",
]

[project.optional-dependencies]
dev = [
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "isort>=5.13.0"
]

[dependency-groups]
dev = [
    "playwright>=1.55.0",
    "pytest-mock>=3.15.1",
    "pytest-playwright>=0.7.1",
]

[tool.pytest.ini_options]
markers = [
    "core_suite: curated regression tests that must remain under 100 collected cases",
    "extended: optional coverage executed outside the core suite",
    "no_api_key: tests covering missing API key scenarios",
]
addopts = "-m core_suite"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | eleventools
)/
'''

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["B011"]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["eleventools/*", "*.egg-info/*"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "eleventools/",
    "build/",
    "dist/",
    "*.egg-info/",
]

[tool.bandit]
exclude_dirs = ["tests", "eleventools"]
skips = ["B101", "B601"]
